# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Bora Jogar API
  version: 0.1.0
paths:
...
  /api/v1/venues:
    get:
      summary: List approved venues
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: query, name: city, schema: { type: string } }
        - { in: query, name: latitude, schema: { type: number, format: double, minimum: -90, maximum: 90 } }
        - { in: query, name: longitude, schema: { type: number, format: double, minimum: -180, maximum: 180 } }
      responses: { '200': { description: Approved venues, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Venue' } } } } }, '422': { $ref: '#/components/responses/ValidationError' } }
...
  /api/v1/venues/suggestions:
    post:
      summary: Suggest a venue for approval
      security: [{ cookieAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/VenueSuggestionInput' } } } }
      responses: { '201': { description: Pending venue suggestion }, '422': { $ref: '#/components/responses/ValidationError' } }
  /api/v1/me/preferred-areas:
    get: { summary: List current user's preferred areas, security: [{ cookieAuth: [] }], responses: { '200': { description: Preferred areas } } }
    post:
      summary: Create preferred area
      security: [{ cookieAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PreferredAreaInput' } } } }
      responses: { '201': { description: Preferred area created }, '409': { description: Area limit reached }, '422': { $ref: '#/components/responses/ValidationError' } }
...
  /api/v1/me/favorite-venues:
    get: { summary: List favorite venues, security: [{ cookieAuth: [] }], responses: { '200': { description: Favorite venues } } }
...
  /api/v1/admin/venues/suggestions:
    get: { summary: List pending venue suggestions, security: [{ cookieAuth: [] }], responses: { '200': { description: Pending suggestions }, '403': { $ref: '#/components/responses/Forbidden' } } }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: borajogar_session }
  schemas:
...
    Venue:
      type: object
      required: [id, name, city, latitude, longitude, lightingStatus, surfaceType, accessType, active]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        addressLabel: { type: string, nullable: true }
        city: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        lightingStatus: { type: string, enum: [unknown, no_lighting, has_lighting] }
        surfaceType: { type: string, enum: [unknown, sand, grass, hard_court, other] }
        accessType: { type: string, enum: [public, private, paid_entry, unknown] }
        active: { type: boolean }
        distanceMeters: { type: number, format: double, nullable: true }
    PreferredAreaInput:
      type: object
      required: [label, latitude, longitude, radiusMeters]
      properties:
        label: { type: string, minLength: 1, maxLength: 120 }
        latitude: { type: number, format: double, minimum: -90, maximum: 90 }
        longitude: { type: number, format: double, minimum: -180, maximum: 180 }
        radiusMeters: { type: integer, minimum: 500, maximum: 25000 }
        priority: { type: integer, minimum: 0 }
    VenueSuggestionInput:
      type: object
      required: [name, city, latitude, longitude]
      properties:
        name: { type: string, minLength: 2, maxLength: 160 }
        description: { type: string, nullable: true }
        addressLabel: { type: string, nullable: true }
        city: { type: string, minLength: 1, maxLength: 120 }
        latitude: { type: number, format: double, minimum: -90, maximum: 90 }
        longitude: { type: number, format: double, minimum: -180, maximum: 180 }
...
    ApiError:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message, fields]
          properties:
            code: { type: string }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
        requestId: { type: string }
  responses:
...
    Forbidden: { description: Administrator access required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
...
    ValidationError: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }